+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/dataformats/wkw/WKWDataFormatHelper.scala
@@ -32,34 +32,4 @@ trait WKWDataFormatHelper {
     }
   }
 
-  def voxelTypeToElementClass(voxelType: VoxelType.Value, voxelSize: Int): Box[ElementClass.Value] =
-    (voxelType, voxelSize) match {
-      case (VoxelType.UInt8, 1)  => Full(ElementClass.uint8)
-      case (VoxelType.UInt16, 2) => Full(ElementClass.uint16)
-      case (VoxelType.UInt8, 3)  => Full(ElementClass.uint24)
-      case (VoxelType.UInt32, 4) => Full(ElementClass.uint32)
-      case (VoxelType.UInt64, 8) => Full(ElementClass.uint64)
-      case (VoxelType.Float, 4)  => Full(ElementClass.float)
-      case (VoxelType.Double, 8) => Full(ElementClass.double)
-      case (VoxelType.Int8, 1)   => Full(ElementClass.int8)
-      case (VoxelType.Int16, 2)  => Full(ElementClass.int16)
-      case (VoxelType.Int32, 4)  => Full(ElementClass.int32)
-      case (VoxelType.Int64, 8)  => Full(ElementClass.int64)
-      case _                     => Failure("VoxelType is not supported.")
-    }
-
-  def elementClassToVoxelType(elementClass: ElementClass.Value): (VoxelType.Value, Int) =
-    elementClass match {
-      case ElementClass.uint8  => (VoxelType.UInt8, 1)
-      case ElementClass.uint16 => (VoxelType.UInt16, 1)
-      case ElementClass.uint24 => (VoxelType.UInt8, 3)
-      case ElementClass.uint32 => (VoxelType.UInt32, 1)
-      case ElementClass.uint64 => (VoxelType.UInt64, 1)
-      case ElementClass.float  => (VoxelType.Float, 1)
-      case ElementClass.double => (VoxelType.Double, 1)
-      case ElementClass.int8   => (VoxelType.Int8, 1)
-      case ElementClass.int16  => (VoxelType.Int16, 1)
-      case ElementClass.int32  => (VoxelType.Int32, 1)
-      case ElementClass.int64  => (VoxelType.Int64, 1)
-    }
 }
