+++ webknossos-datastore/app/com/scalableminds/webknossos/datastore/dataformats/wkw/WKWHeader.scala
@@ -28,37 +28,6 @@ object ChunkType extends Enumeration(1) {
   def isCompressed(blockType: ChunkType.Value): Boolean = blockType == LZ4 || blockType == LZ4HC
 }
 
-object VoxelType extends Enumeration(1) {
-  val UInt8, UInt16, UInt32, UInt64, Float, Double, Int8, Int16, Int32, Int64 = Value
-
-  def bytesPerVoxelPerChannel(voxelType: VoxelType.Value): Int = voxelType match {
-    case UInt8  => 1
-    case UInt16 => 2
-    case UInt32 => 4
-    case UInt64 => 8
-    case Float  => 4
-    case Double => 8
-    case Int8   => 1
-    case Int16  => 2
-    case Int32  => 4
-    case Int64  => 8
-  }
-
-  def toArrayDataType(voxelType: VoxelType.Value): ArrayDataType =
-    voxelType match {
-      case UInt8  => ArrayDataType.u1
-      case UInt16 => ArrayDataType.u2
-      case UInt32 => ArrayDataType.u4
-      case UInt64 => ArrayDataType.u8
-      case Float  => ArrayDataType.f4
-      case Double => ArrayDataType.f8
-      case Int8   => ArrayDataType.i1
-      case Int16  => ArrayDataType.i2
-      case Int32  => ArrayDataType.i4
-      case Int64  => ArrayDataType.i8
-    }
-}
-
 case class WKWHeader(
     version: Int,
     numChunksPerShardDimension: Int,
